bam/bai download, size/md5sum comparison are in place now

+++ s3objectqc/bam_qc/downloading.py
@@ -5,6 +5,7 @@ import xmltodict
 import time
 import subprocess
 import calendar
+import hashlib
 from ..job_tracker import move_to_next_step, get_job_json, save_job_json
 
 
@@ -18,36 +19,80 @@ def get_name():
     return name
 
 
-def download_bam_and_get_info(job_dir, object_id, file_name):
+def download_file_and_get_info(job_dir, object_id, file_name):
     global s3_bucket_url
 
-    command =   'cd {} && '.format(job_dir) + \
-                'aws s3 cp ' + s3_bucket_url + \
-                object_id + ' ' + \
-                file_name
+    file_info = {
+        'file_size': None,
+        'file_md5sum': None
+    }
 
-    process = subprocess.Popen(
-            command,
-            shell=True,
-            stdout=subprocess.PIPE,
-            stderr=subprocess.PIPE
-        )
+    fpath = os.path.join(job_dir, file_name)
 
-    out, err = process.communicate()
-    if process.returncode:
-        # should not exit for just this error, improve it later
-        sys.exit('Unable to download xml file.\nError message: {}'.format(err))
-    else:
-        file_info = {
-            'file_size': None,
-            'file_md5sum': None
-        }
-        # file is here: os.path.join(job_dir, file_name)
+    # Only download when file does not already exist.
+    # - This is meant more for repeative testing/debugging without
+    #   having to download large file over and over again.
+    # - In real world, shouldn't have as each time a new run dir is created 
+    if not os.path.isfile(fpath):
+        command =   'cd {} && '.format(job_dir) + \
+                    'aws s3 cp ' + s3_bucket_url + \
+                    object_id + ' ' + \
+                    file_name
 
-        return file_info
+        process = subprocess.Popen(
+                command,
+                shell=True,
+                stdout=subprocess.PIPE,
+                stderr=subprocess.PIPE
+            )
+
+        out, err = process.communicate()
+
+        if process.returncode:
+            # should not exit for just this error, improve it later
+            sys.exit('Unable to download xml file.\nError message: {}'.format(err))
+
+    # TODO: get file size and md5sum
+    file_info['file_size'] = os.path.getsize(fpath)
+    file_info['file_md5sum'] = get_md5(fpath)
+
+    return file_info
+
+
+def get_md5(fname):
+    hash = hashlib.md5()
+    with open(fname) as f:
+        for chunk in iter(lambda: f.read(4096), ""):
+            hash.update(chunk)
+    return hash.hexdigest()
 
 
 def compare_file(job):
+    # we'd like to do
+    # - download bai, check size and md5sum
+    # - download bam, check size and md5sum
+
+    for f in ['bai_file', 'bam_file']:
+        job_dir = job.job_dir
+        object_id = job.job_json.get(f).get('object_id')
+        file_name = job.job_json.get(f).get('file_name')
+        file_info = download_file_and_get_info(job_dir, object_id, file_name)
+
+        mismatch = False
+        if not file_info.get('file_size') == job.job_json.get(f).get('file_size'):
+            job.job_json.get('_runs_').get(job.conf.get('run_id')).get(get_name()).update({
+                f + '-size-mismatch': file_info.get('file_size')
+            })
+            mismatch = True
+
+        if not file_info.get('file_md5sum') == job.job_json.get(f).get('file_md5sum'):
+            job.job_json.get('_runs_').get(job.conf.get('run_id')).get(get_name()).update({
+                f + '-md5sum-mismatch': file_info.get('file_md5sum')
+            })
+            mismatch = True
+
+        if mismatch: return False
+
     return True
